requirements.txt
streamlit
openai
pdfplumber
python-dotenv
fpdf2
camelot-py[cv]
